limit ec2 and gce tags to the proper tagged builds (for puppy and dogstatsd)
